refactor: extract contract address logic into custom hooks (Phase 1)
Extract business logic from LinkContractAddressButton into three focused hooks:
- useContractAddressPairs: manages network/address pair state
- useContractAddressValidation: handles validation with error state
- useContractAddressSave: manages save operations with integrated validation

This reduces the component from ~416 to ~248 lines (~40% reduction) while
improving testability, reusability, and maintainability.
